// src/app/global/models/api_placer.model.ts
export interface API_PLACER {
    placerDumpId: string;
    placerId: string;
    placerName: string;
    hostId: string;
    hostName: string;
    address: string;
    hostCity: string;
    hostState: string;
    postalCode: string;
    averageDwellTime: number;
    footTraffic: number;
    publicationDate: string;
    month: string;
    sourceFile: string;
    dateUploaded: string;
    uploadedBy: string;
}